pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-scrapers"
version = "0.1.0"
description = "Modular web scraping system with GitHub Actions automation"
requires-python = ">=3.11"
dependencies = [
    "firecrawl-py",
    "beautifulsoup4",
    "lxml",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 